:lipstick: adicionando modificacoes finais para o design de services
