+++ admin/src/main/java/com/ezschedule/admin/presenter/adapter/ServicesAdapter.kt
@@ -5,21 +5,13 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.ezschedule.admin.R
 import com.ezschedule.admin.databinding.AdapterServiceBinding
 import com.ezschedule.network.domain.presentation.ServicePresentation
-import com.ezschedule.network.domain.presentation.TenantServicePresentation
 
 class ServicesAdapter(
-    private val services: List<ServicePresentation> = emptyList(),
-    private val tenants: List<TenantServicePresentation> = emptyList(),
-    private val context: Context,
-    val itemClicked: ((Int) -> Unit?)? = null
+    private val services: List<ServicePresentation>,
+    private val context: Context
 ) : RecyclerView.Adapter<ServicesAdapter.ViewHolder>() {
-
-
-    private var selectedPos = RecyclerView.NO_POSITION
-    private var viewHolderList = ArrayList<ViewHolder>()
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val binding =
@@ -28,46 +20,22 @@ class ServicesAdapter(
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        viewHolderList.add(holder)
         holder.bind()
     }
 
-    override fun getItemCount(): Int {
-        if (tenants.isNotEmpty()) {
-            return tenants.size
-        }
-        return services.size
-    }
+    override fun getItemCount(): Int = services.size
 
     inner class ViewHolder(private val binding: AdapterServiceBinding) :
         RecyclerView.ViewHolder(binding.root) {
 
         fun bind() = with(binding) {
-            if (tenants.isNotEmpty()) {
-                tvServiceName.text = tenantsData().name
-                tvServiceUser.text = tenantsData().phoneNumber
-                Glide.with(context)
-                    .load(tenantsData().image)
-                    .into(ivService)
-                itemView.setOnClickListener {
-                    if (selectedPos >= 0) {
-                        viewHolderList[selectedPos].itemView.setBackgroundResource(R.drawable.item_rounded)
-                    }
-                    selectedPos = adapterPosition
-                    itemView.setBackgroundResource(R.drawable.item_rounded_secondary)
-                    itemClicked?.let { it(selectedPos) }
-                }
-            } else {
-                tvServiceName.text = serviceData().name
-                tvServiceUser.text = serviceData().userName
-                Glide.with(context)
-                    .load(serviceData().image)
-                    .into(ivService)
-            }
+            tvServiceName.text = serviceData().name
+            tvServiceUser.text = serviceData().userName
+            Glide.with(context)
+                .load(serviceData().image)
+                .into(ivService)
         }
 
         private fun serviceData() = services[adapterPosition]
-
-        private fun tenantsData() = tenants[adapterPosition]
     }
 }
