+++ admin/src/main/res/layout/fragment_services.xml
@@ -6,6 +6,7 @@
     android:layout_height="match_parent">
 
     <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/frag_cl_service"
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
@@ -35,16 +36,16 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"/>
+            app:layout_constraintTop_toTopOf="parent" />
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/frag_rv_service"
             android:layout_width="match_parent"
-            android:visibility="invisible"
             android:layout_height="@dimen/spacing_none"
             android:layout_marginHorizontal="@dimen/spacing_large"
             android:layout_marginTop="@dimen/spacing_large"
             android:clipToPadding="false"
+            android:visibility="invisible"
             app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
             app:layout_constraintBottom_toTopOf="@id/frag_btn_service"
             app:layout_constraintEnd_toEndOf="parent"
@@ -58,15 +59,20 @@
             android:layout_width="0dp"
             android:layout_height="48dp"
             android:layout_margin="@dimen/spacing_large"
-            app:layout_constraintTop_toBottomOf="@id/frag_rv_service"
             android:text="@string/frag_services_btn_plus"
             android:textAppearance="@style/ButtonAppearance"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/frag_rv_service"
             app:layout_constraintWidth_percent="0.4" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
+
+    <include
+        android:id="@+id/include_add_service"
+        layout="@layout/view_add_service"
+        android:visibility="gone"/>
 
     <include
         android:id="@+id/include_loading"
