+++ admin/src/main/res/layout/view_add_service.xml
@@ -1,0 +1,108 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <ImageButton
+        android:id="@+id/frag_services_btn_back"
+        style="@style/ImageButtonEzSchedule"
+        android:layout_width="0dp"
+        android:layout_height="48dp"
+        android:layout_marginStart="@dimen/spacing_small"
+        android:layout_marginTop="@dimen/spacing_small"
+        android:src="@drawable/ic_previous"
+        app:layout_constraintHorizontal_weight="1"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintWidth_percent="0.4" />
+
+    <androidx.appcompat.widget.SearchView
+        android:id="@+id/frag_sv_service"
+        android:layout_width="match_parent"
+        android:layout_height="48dp"
+        android:layout_marginHorizontal="@dimen/spacing_large"
+        android:layout_marginTop="@dimen/spacing_standard"
+        android:background="@drawable/bg_search_view"
+        android:textSize="@dimen/size_small"
+        android:textStyle="bold"
+        app:iconifiedByDefault="false"
+        app:layout_constraintBottom_toTopOf="@id/frag_rv_service"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/frag_services_btn_back"
+        app:queryBackground="@android:color/transparent"
+        app:queryHint="@string/frag_services_sv_hint" />
+
+    <TextView
+        android:id="@+id/tv_no_residents"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="No residents have been found"
+        android:textSize="@dimen/size_small"
+        android:visibility="gone"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/frag_rv_service"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_marginHorizontal="@dimen/spacing_large"
+        android:layout_marginVertical="@dimen/spacing_large"
+        android:visibility="invisible"
+        app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
+        app:layout_constraintBottom_toTopOf="@id/frag_til_name_new_service"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/frag_sv_service"
+        tools:listitem="@layout/adapter_service" />
+
+    <com.google.android.material.textfield.TextInputLayout
+        android:id="@+id/frag_til_name_new_service"
+        style="@style/Widget.MaterialComponents.TextInputLayout.FilledBox"
+        android:layout_width="0dp"
+        android:layout_height="48dp"
+        android:layout_marginEnd="16dp"
+        android:layout_marginBottom="@dimen/spacing_small"
+        app:boxCornerRadiusBottomEnd="@dimen/corners_tiny"
+        app:boxCornerRadiusBottomStart="@dimen/corners_tiny"
+        app:boxCornerRadiusTopEnd="@dimen/corners_tiny"
+        app:boxCornerRadiusTopStart="@dimen/corners_tiny"
+        app:boxStrokeWidth="0dp"
+        app:boxStrokeWidthFocused="0dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toStartOf="@id/btn_add_service"
+        app:layout_constraintStart_toStartOf="@id/frag_rv_service"
+        app:layout_constraintTop_toBottomOf="@id/frag_rv_service">
+
+        <EditText
+            android:id="@+id/edt_name_new_service"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:hint="@string/frag_services_et_hint"
+            android:inputType="text"
+            android:paddingLeft="@dimen/spacing_small"
+            android:textSize="@dimen/size_small" />
+
+    </com.google.android.material.textfield.TextInputLayout>
+
+    <Button
+        android:id="@+id/btn_add_service"
+        style="@style/ButtonEzSchedule"
+        android:layout_width="0dp"
+        android:layout_height="48dp"
+        android:text="@string/frag_services_btn_desc"
+        android:textAppearance="@style/ButtonAppearance"
+        app:layout_constraintBottom_toBottomOf="@id/frag_til_name_new_service"
+        app:layout_constraintEnd_toEndOf="@id/frag_rv_service"
+        app:layout_constraintStart_toEndOf="@id/frag_til_name_new_service"
+        app:layout_constraintTop_toTopOf="@id/frag_til_name_new_service"
+        app:layout_constraintWidth_percent=".3" />
+</androidx.constraintlayout.widget.ConstraintLayout>
+
