--- admin/src/main/res/layout/view_service_bottom_sheet.xml
@@ -1,128 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="@drawable/bg_bottom_sheet_ezschedule"
-    app:layout_behavior="@string/bottom_sheet_behavior">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <ImageView
-            android:id="@+id/iv_lower"
-            android:layout_width="100dp"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="@dimen/spacing_standard"
-            android:scaleType="centerCrop"
-            android:src="@drawable/lower"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:ignore="ContentDescription" />
-
-        <TextView
-            android:id="@+id/tv_title"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="@dimen/spacing_large"
-            android:gravity="center"
-            android:text="@string/frag_tv_title"
-            android:textAppearance="@style/TitleAppearance"
-            android:textColor="@color/colorVariant"
-            app:layout_constraintBottom_toTopOf="@id/frag_sv_service"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/iv_lower" />
-
-        <androidx.appcompat.widget.SearchView
-            android:id="@+id/frag_sv_service"
-            android:layout_width="match_parent"
-            android:layout_height="48dp"
-            android:layout_marginHorizontal="@dimen/spacing_large"
-            android:layout_marginTop="@dimen/spacing_standard"
-            android:background="@drawable/bg_search_view"
-            android:textSize="@dimen/size_small"
-            android:textStyle="bold"
-            app:iconifiedByDefault="false"
-            app:layout_constraintBottom_toTopOf="@id/frag_rv_service"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_title"
-            app:queryBackground="@android:color/transparent"
-            app:queryHint="@string/frag_services_sv_hint" />
-
-        <TextView
-            android:id="@+id/tv_no_residents"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="No residents have been found"
-            android:textSize="@dimen/size_small"
-            android:visibility="gone"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/frag_rv_service"
-            android:layout_width="match_parent"
-            android:layout_height="320dp"
-            android:layout_marginHorizontal="@dimen/spacing_large"
-            android:layout_marginVertical="@dimen/spacing_large"
-            android:visibility="invisible"
-            app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-            app:layout_constraintBottom_toTopOf="@id/frag_til_name_new_service"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/frag_sv_service"
-            tools:listitem="@layout/adapter_service" />
-
-        <com.google.android.material.textfield.TextInputLayout
-            android:id="@+id/frag_til_name_new_service"
-            style="@style/Widget.MaterialComponents.TextInputLayout.FilledBox"
-            android:layout_width="0dp"
-            android:layout_height="48dp"
-            android:layout_marginEnd="16dp"
-            android:layout_marginBottom="@dimen/spacing_small"
-
-            app:boxCornerRadiusBottomEnd="@dimen/corners_tiny"
-            app:boxCornerRadiusBottomStart="@dimen/corners_tiny"
-            app:boxCornerRadiusTopEnd="@dimen/corners_tiny"
-            app:boxCornerRadiusTopStart="@dimen/corners_tiny"
-            app:boxStrokeWidth="0dp"
-            app:boxStrokeWidthFocused="0dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toStartOf="@id/btn_add_service"
-            app:layout_constraintStart_toStartOf="@id/frag_rv_service"
-            app:layout_constraintTop_toBottomOf="@id/frag_rv_service">
-
-            <EditText
-                android:id="@+id/edt_name_new_service"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:hint="@string/frag_services_et_hint"
-                android:inputType="text"
-                android:paddingLeft="@dimen/spacing_small"
-                android:textSize="@dimen/size_small" />
-
-        </com.google.android.material.textfield.TextInputLayout>
-
-        <Button
-            android:id="@+id/btn_add_service"
-            style="@style/ButtonEzSchedule"
-            android:layout_width="0dp"
-            android:layout_height="48dp"
-            android:text="@string/frag_services_btn_desc"
-            android:textAppearance="@style/ButtonAppearance"
-            app:layout_constraintBottom_toBottomOf="@id/frag_til_name_new_service"
-            app:layout_constraintEnd_toEndOf="@id/frag_rv_service"
-            app:layout_constraintStart_toEndOf="@id/frag_til_name_new_service"
-            app:layout_constraintTop_toTopOf="@id/frag_til_name_new_service"
-            app:layout_constraintWidth_percent=".3" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</ScrollView>
